changed page_size to uint32_t ints

+++ inc/page_layout.h
@@ -15,24 +15,24 @@
 // insert and delete operation return 1 upon success, else they return 0 for failure
 
 // to insert a tuple at the given index in the given page
-int insert_tuple(void* page, uint64_t page_size, const tuple_def* tpl_d, uint16_t index, const void* external_tuple);
+int insert_tuple(void* page, uint32_t page_size, const tuple_def* tpl_d, uint16_t index, const void* external_tuple);
 
 // to remove a tuple at the given index in the page
-int delete_tuple(void* page, uint64_t page_size, const tuple_def* tpl_d, uint16_t index);
+int delete_tuple(void* page, uint32_t page_size, const tuple_def* tpl_d, uint16_t index);
 
 
 
 
 // to check if a tuple at the given index in the page has been deleted
 // 1 means the tuple is deleted OR it does not exists
-int is_deleted_tuple(const void* page, uint64_t page_size, const tuple_def* tpl_d, uint16_t index);
+int is_deleted_tuple(const void* page, uint32_t page_size, const tuple_def* tpl_d, uint16_t index);
 
 
 
 
 // returns the index that will be assigned for a new tuple upon insertion
 // i.e. an iterator in the for loop must be lesser than the return value of this function
-uint16_t get_tuple_count(const void* page, uint64_t page_size, const tuple_def* tpl_d);
+uint16_t get_tuple_count(const void* page, uint32_t page_size, const tuple_def* tpl_d);
 
 
 
@@ -42,17 +42,17 @@ uint16_t get_tuple_count(const void* page, uint64_t page_size, const tuple_def* 
 // index attribute must be lesser than get_index_for_new_tuple(), else you get NULL
 
 // returns pointer to nth tuple in the page
-void* seek_to_nth_tuple(const void* page, uint64_t page_size, const tuple_def* tpl_d, uint16_t index);
+void* seek_to_nth_tuple(const void* page, uint32_t page_size, const tuple_def* tpl_d, uint16_t index);
 
 
 
 
 // returns true if the given page has enough space to accomodate the given tuple
-int can_accomodate_tuple(const void* page, uint64_t page_size, const tuple_def* tpl_d, const void* external_tuple);
+int can_accomodate_tuple(const void* page, uint32_t page_size, const tuple_def* tpl_d, const void* external_tuple);
 
 // returns total free space inside a given page
 // this is the space that will be freed if the page was to be compacted thoroughly
-uint64_t get_free_space(const void* page, uint64_t page_size, const tuple_def* tpl_d);
+uint32_t get_free_space(const void* page, uint32_t page_size, const tuple_def* tpl_d);
 
 
 
@@ -62,13 +62,13 @@ uint64_t get_free_space(const void* page, uint64_t page_size, const tuple_def* t
 //    * reducing the extra area given to the VARIABLE_SIZED elements in the tuple
 //    * removing tomb stones for deleted records and their records space in both VARIABLE_SIZED and fixed sized tuples
 // returns 1, if any of the compaction operation was performed, else 0
-int compact_page(const void* page, uint64_t page_size, const tuple_def* tpl_d);
+int compact_page(const void* page, uint32_t page_size, const tuple_def* tpl_d);
 
 
 
 
 // prints all tuples in the page including the deleted ones with their tomb stones
-void print_all_tuples(const void* page, uint64_t page_size, const tuple_def* tpl_d);
+void print_all_tuples(const void* page, uint32_t page_size, const tuple_def* tpl_d);
 
 #endif
